esm - convert userDataPath.js to userDataPath.ts
